fix: fix code and configs

+++ src/main/java/edu/sharif/simple/SimpleRepository.java
@@ -2,5 +2,5 @@ package edu.sharif.simple;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SimpleRepository extends JpaRepository<SimpleEntity, Long> {
+public interface SimpleRepository extends JpaRepository<SimpleEntity, Integer> {
 }
